References/Web prediksi rumah/requirements.txt
click==8.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
Jinja2
jsonify==0.5
numpy==1.19.5
pandas==1.3.3
scikit-image==0.18.1
scikit-learn
Scrapy==2.5.1
urllib3==1.26.7
w3lib==1.22.0
Werkzeug==2.0.2
